spp_nfv: remove global variable client_id
`client_id` is defined as a global variable in `params.h`, but it is
better to use get_client_id() instead. This update is to replace it.
This update is also to make a change for formatter function
`get_sec_stats_json()` which takes client ID as uint16_t, but
get_client_id() returns the ID as int.

+++ src/nfv/nfv_status.h
@@ -6,7 +6,7 @@
 #define _NFV_STATUS_H_
 
 /* Get status of spp_nfv or spp_vm as JSON format. */
-void get_sec_stats_json(char *str, uint16_t client_id,
+void get_sec_stats_json(char *str, int client_id,
 		const char *running_stat,
 		uint8_t lcore_id_used[RTE_MAX_LCORE],
 		struct port *ports_fwd_array,
